refactor: move TimeRange into models to break core-to-cli dependency
