+++ src/models/mod.rs
@@ -11,6 +11,7 @@ pub mod analysis;
 pub mod claude;
 pub mod codex;
 pub mod copilot;
+pub mod filter;
 pub mod gemini;
 pub mod grok;
 pub mod provider;
@@ -21,6 +22,7 @@ pub use self::analysis::*;
 pub use self::claude::*;
 pub use self::codex::*;
 pub use self::copilot::*;
+pub use self::filter::*;
 pub use self::gemini::*;
 pub use self::grok::*;
 pub use self::provider::*;
