+++ src/session/opencode.rs
@@ -20,7 +20,7 @@
 //! sessions that switch model mid-stream are split before aggregation.
 
 use crate::VERSION;
-use crate::cli::TimeRange;
+use crate::models::TimeRange;
 use crate::constants::FastHashMap;
 use crate::models::{CodeAnalysis, CodeAnalysisRecord, ExtensionType};
 use crate::session::diagnostics::{
